+++ sdk/storage/azure-storage-blobs/test/blob_container_client_test.cpp
@@ -468,7 +468,6 @@ namespace Azure { namespace Storage { namespace Test {
     auto brokenLease = *leaseClient.Break();
     EXPECT_TRUE(brokenLease.ETag.HasValue());
     EXPECT_TRUE(IsValidTime(brokenLease.LastModified));
-    EXPECT_EQ(brokenLease.LeaseTime, 0);
 
     leaseClient
         = Blobs::BlobLeaseClient(containerClient, Blobs::BlobLeaseClient::CreateUniqueLeaseId());
@@ -476,7 +475,6 @@ namespace Azure { namespace Storage { namespace Test {
     brokenLease = *leaseClient.Break();
     EXPECT_TRUE(brokenLease.ETag.HasValue());
     EXPECT_TRUE(IsValidTime(brokenLease.LastModified));
-    EXPECT_NE(brokenLease.LeaseTime, 0);
 
     Blobs::BreakBlobLeaseOptions options;
     options.BreakPeriod = std::chrono::seconds(0);